Improve code organisation
- split single code file into smaller units grouped by semantic
- add lots of Kdoc documentation
- rename for better consistency
